Change Value Type from Int to Int32 for Profile ID's

+++ Apple Project/Cities.swift
@@ -12,12 +12,12 @@ import CoreLocation
 
 class Cities {
     
-    let profileID: Double
+    let profileID: Int32
     let coordinates: CLLocationCoordinate2D
     let country: String
     let city: String
     
-    init(profileID: Double, coordinates: CLLocationCoordinate2D, country: String, city: String) {
+    init(profileID: Int32, coordinates: CLLocationCoordinate2D, country: String, city: String) {
         self.profileID = profileID
         self.coordinates = coordinates
         self.country = country
